refactor: restructure MarketDetail and MarketHome components into V1 versions
- Moved existing MarketDetail and MarketHome components to new V1 subdirectories.
- Updated exports to use the new V1 components, ensuring backward compatibility.
- Simplified the structure for better maintainability and future enhancements.
