+++ packages/kit/src/views/Market/MarketDetailV1/index.ts
@@ -1,0 +1,1 @@
+export { default as MarketDetailV1 } from './MarketDetail';
