+++ packages/kit/src/views/Market/MarketHomeV1/index.ts
@@ -1,0 +1,1 @@
+export { default as MarketHomeV1 } from './MarketHome';
